+++ application/models/Flight.php
@@ -8,6 +8,8 @@
  */
 class Flight extends CSV_Model
 {
+    
+    private $search;
 
 	// Constructor
 	//load flights.csv
@@ -110,6 +112,7 @@ class Flight extends CSV_Model
 				}
 			}
 		}
+        $this->search = $searchresult;
 		return $searchresult;
 	}
 }