+++ changelog.md
@@ -5,7 +5,9 @@ Changelog
 - added unit test for single flights
 - added unit test for business rule conformation
 - added unit test for Fleet model
+- added unit test for Flight model
 - added unit test for plane/server verification
+
 
 ### #6 2017-11-12 - Gabriel
 - added Booking Search on welcome page